Move traced_impl to trace.traced_impl
